update to logging mechanism
