+++ data/cot_responses/logs/.gitignore
