+++ scripts/putnamlike2_split_cots.py
@@ -1,76 +1,12 @@
 #!/usr/bin/env python3
-"""E.g.
-
-From **/chainscope/** run:
-
-```
-python3 -m dotenv run python3 \
-  scripts/putnam/putnamlike2_split_cots.py \
-  d/cot_responses/instr-v0/default_sampling_params/filtered_putnambench/google__gemini-exp-1206:free_v0_prefix_1_just_correct_responses.yaml \
-  --model_id "openai/gpt-4o" \
-  --max_retries=3 \
-  --prefix=1 \
-  --max_new_tokens_override=100000 \
-  --verbose
-```
-
-"""
 
 import logging
-import datetime
 from pathlib import Path
 import click
 
-from chainscope.cot_splitting import split_cot_responses
-from chainscope.typing import *
-
-
-def setup_logging(verbose: bool, script_name: str) -> str:
-    """Set up logging to both console and file.
-    
-    Args:
-        verbose: Whether to log at INFO level (True) or WARNING level (False)
-        script_name: Name of the script for the log filename
-    
-    Returns:
-        Path to the log file
-    """
-    # Create logs directory if it doesn't exist
-    logs_dir = Path("logs")
-    logs_dir.mkdir(exist_ok=True)
-    
-    # Create log filename with timestamp
-    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-    log_filename = f"{script_name}_{timestamp}.log"
-    log_path = logs_dir / log_filename
-    
-    # Set up logging
-    log_level = logging.INFO if verbose else logging.WARNING
-    
-    # Configure root logger
-    root_logger = logging.getLogger()
-    root_logger.setLevel(log_level)
-    
-    # Clear existing handlers
-    for handler in root_logger.handlers[:]:
-        root_logger.removeHandler(handler)
-    
-    # Create console handler
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(log_level)
-    console_format = logging.Formatter('%(levelname)s: %(message)s')
-    console_handler.setFormatter(console_format)
-    root_logger.addHandler(console_handler)
-    
-    # Create file handler
-    file_handler = logging.FileHandler(log_path)
-    file_handler.setLevel(log_level)
-    file_format = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    file_handler.setFormatter(file_format)
-    root_logger.addHandler(file_handler)
-    
-    logging.info(f"Logging to {log_path}")
-    return str(log_path)
+from src.cot_splitting import split_cot_responses
+from src.typing import *
+from src.utils import setup_logging
 
 
 @click.command()
