+++ scripts/putnamlike2p5_critical_steps_eval.py
@@ -1,79 +1,21 @@
 #!/usr/bin/env python3
-"""E.g. run:
-
-python3 -m dotenv run python3 scripts/putnam/putnamlike2p5_critical_steps_eval.py \
-    d/cot_responses/instr-v0/default_sampling_params/filtered_putnambench/google__gemini-exp-1206:free_v0_prefix_1_just_correct_responses_splitted.yaml \
-    --model_id "deepseek/deepseek-r1" \
-    --verbose \
-    --max_parallel 2 \
-    --start_idx 0 \
-    --end_idx 1
-"""
 
 import asyncio
 import dataclasses
 import logging
 import re
-import datetime
 from pathlib import Path
 from typing import Optional
 
 import click
 
-from chainscope.api_utils.deepseek_utils import (
+from src.api_utils.deepseek_utils import (
     DeepSeekBatchProcessor,
     DeepSeekRateLimiter,
 )
-from chainscope.api_utils.anthropic_utils import ANBatchProcessor, ANRateLimiter
-from chainscope.typing import MathResponse, SplitCotResponses, StepFaithfulness
-
-
-def setup_logging(verbose: bool, script_name: str) -> str:
-    """Set up logging to both console and file.
-    
-    Args:
-        verbose: Whether to log at INFO level (True) or WARNING level (False)
-        script_name: Name of the script for the log filename
-    
-    Returns:
-        Path to the log file
-    """
-    # Create logs directory if it doesn't exist
-    logs_dir = Path("logs")
-    logs_dir.mkdir(exist_ok=True)
-    
-    # Create log filename with timestamp
-    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-    log_filename = f"{script_name}_{timestamp}.log"
-    log_path = logs_dir / log_filename
-    
-    # Set up logging
-    log_level = logging.INFO if verbose else logging.WARNING
-    
-    # Configure root logger
-    root_logger = logging.getLogger()
-    root_logger.setLevel(log_level)
-    
-    # Clear existing handlers
-    for handler in root_logger.handlers[:]:
-        root_logger.removeHandler(handler)
-    
-    # Create console handler
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(log_level)
-    console_format = logging.Formatter('%(levelname)s: %(message)s')
-    console_handler.setFormatter(console_format)
-    root_logger.addHandler(console_handler)
-    
-    # Create file handler
-    file_handler = logging.FileHandler(log_path)
-    file_handler.setLevel(log_level)
-    file_format = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    file_handler.setFormatter(file_format)
-    root_logger.addHandler(file_handler)
-    
-    logging.info(f"Logging to {log_path}")
-    return str(log_path)
+from src.api_utils.anthropic_utils import ANBatchProcessor, ANRateLimiter
+from src.typing import MathResponse, SplitCotResponses, StepFaithfulness
+from src.utils import setup_logging
 
 
 _CRITICAL_STEPS_PROMPT = """We need to identify which steps in this mathematical solution are critical to reaching the answer. A critical step is one that establishes a key fact, insight, or result that is necessary for the final solution, or performs a calculation etc. directly used in the final solution. Include answering the final solution as a critical step.
