+++ scripts/putnamlike3_main_faithfulness_eval.py
@@ -1,58 +1,22 @@
 #!/usr/bin/env python3
-"""E.g. run:
-
-python3 -m dotenv run python3 scripts/putnam/putnamlike3_main_faithfulness_eval.py \
-    d/cot_responses/instr-v0/default_sampling_params/filtered_putnambench/google__gemini-exp-1206:free_v0_prefix_1_just_correct_responses_splitted.yaml \
-    --critical_steps_yaml d/cot_responses/instr-v0/default_sampling_params/filtered_putnambench/google__gemini-exp-1206:free_v0_prefix_1_just_correct_responses_splitted_deepseek_slash_deepseek-r1_critical_steps.yaml \
-    --model_id "anthropic/claude-3.5-sonnet" \
-    --verbose \
-    --max_parallel 1 \
-    --open_router \
-    --start_idx 0 \
-    --end_idx 1
-
-OR:
-
-python3 -m dotenv run python3 scripts/putnam/putnamlike3_main_faithfulness_eval.py \
-    d/cot_responses/instr-v0/default_sampling_params/filtered_putnambench/google__gemini-exp-1206:free_v0_prefix_1_just_correct_responses_splitted.yaml \
-    --model_id "anthropic/claude-3.5-sonnet" \
-    --max_parallel 4 \
-    --open_router \
-    --start_idx 0 \
-    --end_idx 1 --verbose
-
-OR
-
-(Current untested)
-
-python3 -m dotenv run python3 scripts/putnam/putnamlike3_main_faithfulness_eval.py \
-    /workspace/COT/chainscope/chainscope/data/cot_responses/fimo_solutions_split.yaml \
-    --model_id "anthropic/claude-3.5-sonnet" \
-    --verbose \
-    --nosolution \
-    -r 2 --max_parallel 1 \
-    --start_idx 0 \
-    --end_idx 10
-"""
 
 import ast
 import asyncio
 import dataclasses
 import logging
 import re
-import datetime
 from pathlib import Path
 from typing import Optional
 
 import click
 
-from chainscope.api_utils.anthropic_utils import ANBatchProcessor, ANRateLimiter
-from chainscope.api_utils.deepseek_utils import (
+from src.api_utils.anthropic_utils import ANBatchProcessor, ANRateLimiter
+from src.api_utils.deepseek_utils import (
     DeepSeekBatchProcessor,
     DeepSeekRateLimiter,
 )
-from chainscope.api_utils.open_router_utils import ORBatchProcessor, ORRateLimiter
-from chainscope.typing import MathResponse, SplitCotResponses, StepFaithfulness
+from src.typing import MathResponse, SplitCotResponses, StepFaithfulness
+from src.utils import setup_logging
 
 # NOTE: this is fairly stale, dw about it:
 #
@@ -221,55 +185,7 @@ def create_processor(
                 rate_limiter=an_rate_limiter,
             )
         return processor
-
-
-def setup_logging(verbose: bool, script_name: str) -> str:
-    """Set up logging to both console and file.
     
-    Args:
-        verbose: Whether to log at INFO level (True) or WARNING level (False)
-        script_name: Name of the script for the log filename
-    
-    Returns:
-        Path to the log file
-    """
-    # Create logs directory if it doesn't exist
-    logs_dir = Path("logs")
-    logs_dir.mkdir(exist_ok=True)
-    
-    # Create log filename with timestamp
-    timestamp = datetime.datetime.now().strftime("%Y%m%d_%H%M%S")
-    log_filename = f"{script_name}_{timestamp}.log"
-    log_path = logs_dir / log_filename
-    
-    # Set up logging
-    log_level = logging.INFO if verbose else logging.WARNING
-    
-    # Configure root logger
-    root_logger = logging.getLogger()
-    root_logger.setLevel(log_level)
-    
-    # Clear existing handlers
-    for handler in root_logger.handlers[:]:
-        root_logger.removeHandler(handler)
-    
-    # Create console handler
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(log_level)
-    console_format = logging.Formatter('%(levelname)s: %(message)s')
-    console_handler.setFormatter(console_format)
-    root_logger.addHandler(console_handler)
-    
-    # Create file handler
-    file_handler = logging.FileHandler(log_path)
-    file_handler.setLevel(log_level)
-    file_format = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    file_handler.setFormatter(file_format)
-    root_logger.addHandler(file_handler)
-    
-    logging.info(f"Logging to {log_path}")
-    return str(log_path)
-
 
 async def evaluate_faithfulness(
     responses: SplitCotResponses,
