+++ src/api_utils/google_utils.py
@@ -19,6 +19,7 @@ from src.api_utils.batch_processor import (BatchItem, BatchProcessor,
                                                   BatchResult)
 from src.typing import (AnthropicBatchInfo, DatasetParams,
                                QuestionResponseId, SamplingParams)
+from src.utils import setup_logging
 
 load_dotenv()
 
@@ -48,53 +49,6 @@ class AnthropicLimits:
     org_tpm_remaining: int = 80000  # Organization tokens per minute limit
     org_tpm_reset: str = ""  # When the org TPM limit resets
 
-
-def setup_logging(verbose: bool, script_name: str) -> str:
-    """Set up logging to both console and file.
-    
-    Args:
-        verbose: Whether to log at INFO level (True) or WARNING level (False)
-        script_name: Name of the script for the log filename
-    
-    Returns:
-        Path to the log file
-    """
-    # Create logs directory if it doesn't exist
-    logs_dir = Path("logs")
-    logs_dir.mkdir(exist_ok=True)
-    
-    # Create log filename with timestamp
-    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
-    log_filename = f"{script_name}_google_{timestamp}.log"
-    log_path = logs_dir / log_filename
-    
-    # Set up logging
-    log_level = logging.INFO if verbose else logging.WARNING
-    
-    # Configure root logger
-    root_logger = logging.getLogger()
-    root_logger.setLevel(log_level)
-    
-    # Clear existing handlers
-    for handler in root_logger.handlers[:]:
-        root_logger.removeHandler(handler)
-    
-    # Create console handler
-    console_handler = logging.StreamHandler()
-    console_handler.setLevel(log_level)
-    console_format = logging.Formatter('%(levelname)s: %(message)s')
-    console_handler.setFormatter(console_format)
-    root_logger.addHandler(console_handler)
-    
-    # Create file handler
-    file_handler = logging.FileHandler(log_path)
-    file_handler.setLevel(log_level)
-    file_format = logging.Formatter('%(asctime)s - %(name)s - %(levelname)s - %(message)s')
-    file_handler.setFormatter(file_format)
-    root_logger.addHandler(file_handler)
-    
-    logging.info(f"Logging to {log_path}")
-    return str(log_path)
 
 @dataclass
 class GORateLimiter:
